requirements.txt
Django>=3.0,<4.0
psycopg2>=2.9
opencv-python==4.6.0.66




asgiref==3.5.2
# backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
cloudinary==1.29.0
cycler==0.11.0
# Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
fonttools==4.33.3
idna==3.3
# keras==2.9.0
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
# opencv-python==4.6.0.66
opencv-contrib-python-headless==4.6.0.66
# skia-python>=87.4
# opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0